fixed #738: consolidated stats into stat tests
